refactor(core): the orchestration follows the state out of the namespace (#1607)
With the controller and the supersession token owned by `InFlightNavigation`,
`#executeNavigation` / `#finishAsyncNavigation` / `#handleNoGuardsLeave` and the
two-pass prologue around them had no `this` left to need: they are functions
over `(deps, inFlight, plan)` in `transition/executeNavigation.ts`. The
namespace keeps what it actually is — the entry points, their fire-and-forget
checkpoint, and the DI bag: 962 → 397 lines.

Order matters and it is the opposite of the intuitive one: moving methods first
would have threaded the mutable state through parameters, producing more of the
13-position signatures `guardPhase.ts` already pays for deliberately.

Perf-neutral, measured against the production bundle rather than the tsx bench
runner (which distorts per-op closure cost ~100x): alternating OLD/NEW pairs are
bit-identical on allocations (bare 1384 B, guard 1922 B) and inside the A/A
floor on time. Two earlier readings suggested -45 B on OLD and did not
reproduce across four interleaved runs.

Gates re-measured after the move, all three still discriminate: emptying the
`runStep` cancellation check fails 4 tests, dropping `PRE_SUPPRESSED.has` fails
1, allocating a controller unconditionally in the no-guards leave arc fails 2.
Zero test delta (positive control), 3840/432/153 green, coverage 100%.
